⚡ Bolt: Resolve CI failures, duplicate enums, mismatch fields, and non-exhaustive matching
- Resolved multiple duplicate enum/struct declarations in src/security/vulnerability.rs and mocked testing payload assistants.
- Fixed ShellRepl fields mismatch and resolved undeclared types inside src/shell/repl.rs.
- Resolved multiple Package::new and Hash trait bounds for Version in src/sigpkg/mod.rs.
- Completed SimpleCompositor field initialization inside src/graphics/compositor.rs.
- Resolved return type mismatch in ContextWindowPruner::new inside src/ai/orchestrator.rs.
- Cast AtomicUsize loads to u32 before transmuting to target enum types across oci_runtime, runtime, device, driver, and framework.
- Derived Clone and Copy on RuntimeStats inside src/container/runtime.rs to prevent move errors.

+++ src/boot/uefi.rs
@@ -39,7 +39,7 @@ impl SimpleUEFIBootloader {
 }
 
 impl UEFIBootloader for SimpleUEFIBootloader {
-    fn phase(&self) -> BootPhase { unsafe { core::mem::transmute(self.phase.load(Ordering::SeqCst)) } }
+    fn phase(&self) -> BootPhase { unsafe { core::mem::transmute(self.phase.load(Ordering::SeqCst) as u32) } }
     fn load_kernel(&mut self, _kernel_data: &[u8]) -> Result<BootStatus, BootError> {
         self.phase.store(BootPhase::LoadKernel as usize, Ordering::SeqCst);
         self.kernel_loaded.store(1, Ordering::SeqCst);
