+++ src/driver/framework.rs
@@ -75,7 +75,7 @@ impl Driver for SimpleStorageDriver {
         DriverType::Storage
     }
     fn state(&self) -> DriverState {
-        unsafe { core::mem::transmute(self.state.load(Ordering::SeqCst)) }
+        unsafe { core::mem::transmute(self.state.load(Ordering::SeqCst) as u32) }
     }
     fn set_state(&self, state: DriverState) {
         self.state.store(state as usize, Ordering::SeqCst);
